refactor(tests): extract fixture data and mocks into shared helpers module
Moves all inline fixture constants and mock helper functions from
route-ui-sync.fixture.spec.ts into tests/helpers/routeUiSyncFixtures.ts
so they can be reused across test files without duplication.
